Tables-PDF/requirements.txt
fastapi 
uvicorn 
google-cloud-bigquery 
google-cloud-storage